refactor: 폼 처리 로직을 분리하고 이벤트 핸들러를 구성하여 코드 구조 개선
